api/requirements.txt
alembic==1.15.1
annotated-types==0.7.0
anyio==4.8.0
bcrypt==4.3.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.0
email_validator==2.2.0
exceptiongroup==1.2.2
fastapi==0.115.11
greenlet==3.1.1
h11==0.14.0
idna==3.10
Mako==1.3.9
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
requests==2.32.3
rsa==4.9
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starkbank-ecdsa==2.2.0
starlette==0.46.1
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
